print out the first 50Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,17 +8,17 @@
 
 int main(void)
 {
-	int num1 = 1;
-	int num2 = 2;
+	unsigned long int num1 = 1;
+	unsigned long int num2 = 2;
 	int i;
 
-	int next;
+	unsigned long int next;
 
-	printf("%d, ", num1);
+	printf("%lu, ", num1);
 
 	for (i = 0; i < 50; i++)
 	{
-		printf("%d, ", num2);
+		printf("%lu, ", num2);
 		next = num1 + num2;
 
 		num1 = num2;
